backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
gunicorn==21.2.0
boto3==1.34.69
django-storages==1.14.2
python-dotenv==1.0.1
Pillow==10.2.0
whitenoise==6.6.0
